NX-OS tests and improving tests
