+++ tests_new/test_nxos/test_nxos.sh
@@ -7,6 +7,7 @@ echo "Starting tests...good luck:" \
 && cd .. \
 && py.test -v -s -x test_netmiko_show.py --test_device nxos1 \
 && py.test -v -s -x test_netmiko_config.py --test_device nxos1 \
+&& py.test -v -s -x test_netmiko_config_acl.py --test_device nxos1 \
 && py.test -v -s -x test_netmiko_scp.py --test_device nxos1 \
 || RETURN_CODE=1
